+++ blog/api/v1/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import BlogPageAPI
+
+urlpatterns = [
+    path('v1/blog/posts', BlogPageAPI.as_view()),
+]
