+++ blog/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import AddPost, HomePage, AboutPage, BlogPage, PostPage, AuthorPost, SearchPost, BlogPageAPI
+from .views import AddPost, HomePage, AboutPage, BlogPage, PostPage, AuthorPost, SearchPost
 
 app_name = 'blog'
 urlpatterns = [
@@ -10,5 +10,4 @@ urlpatterns = [
     path('blog/new-post/', AddPost.as_view(), name='new_post'),
     path('blog/post/<int:pk>/', PostPage.as_view(), name='post'),
     path('blog/search/', SearchPost.as_view(), name='search'),
-    path('blog/api/v1/posts_list', BlogPageAPI.as_view(), name='posts_api'),
 ]
